Refactored course module modification
In order to get course module modification working again
I had to change the form template and the form script.

I have added a macro variable for the requirement block
of the template in order to reuse it for both, empty
constraint fields and listing all the defined constraints
for the course.

The input field copy script did not work anymore, since
the input fields were not empty anymore, hence I added
a function which does exactly that: reset input fields
with jQuery.

Signed-off-by: Konrad Reiche <konrad.reiche@googlemail.com>