Remove default user selection
In the template for course responsibilities the currently logged
in user was chosen as default selection.

When the site was accessed and courses were assigned to the default
users no changes were made to the database since the user was not
part of the parameters yet. This could have been solved but the
feature did not seem to be convenient anyway.

Signed-off-by: Konrad Reiche <konrad.reiche@googlemail.com>