Style course schedule
I have customized the course schedule in a way that it looks
now more like a proper schedule: table cells belonging to the
same course are merged and highlighted.

For the merging I have declared a counter variable. Each time
a new course is found in the iteration the counter variable is
set to the course duration. Each time another course is found
and the variable is not zero yet, it is decremented.

In addition I made some minor changes to the table styling,
like centering the text both, vertically and horizontally.

Signed-off-by: Konrad Reiche <konrad.reiche@googlemail.com>