Implement scheduler feedback
Instead of using dedicated URLs for displaying the result
of a schedule request I have extended the schedule method
by returning a Feedback object.

This feedback object is initialized during the check for
schedulability of the course data. The same object is used
as a session attribute to display the details of the scheduler
request. If the course data is not schedulable the reasons are
read directly from the Feedback object during the template
processing.

The session object is removed immediately after it was read.
This way the user gets a notification after the scheduler was
executed, but after refreshing the notification will disappear.

By synchronizing on the Scheduler object whenever the session
attribute is written or read thread-safety is guaranteed.

Signed-off-by: Konrad Reiche <konrad.reiche@googlemail.com>