Update course module references
Until now when a course module was removed all entities
referencing the course module's courses were pointing to
null. This resulted in multiple errors when accessing
other servlet modules.

This included the lecturer responsibilities and the course
schedule since both can have references to course entities.
For that reason both reference types are cleaned up when
a course module is removed by removing the references
as well.

In the process I have extended the MongoDB methods by two
delete methods which take an entity object or a query as
input.

I have also added the first link in the scheduler error
message so the user can fix the problems in fewer steps.

Signed-off-by: Konrad Reiche <konrad.reiche@googlemail.com>