Fix comparison bug in responsibilities template
Apparently Freemarker bases its seq_contains method
for Java objects on their toString method and not
on equals, compareTo or hashCode.

In the course responsibilities template this resulted
in lecturers being responsibile for courses they
weren't set for because they are responsible for a
course with the same name and type.

To fix this I have extended the toString method by
the course id.

Signed-off-by: Konrad Reiche <konrad.reiche@googlemail.com>