Add sign up form
A sign up form was added for new users to register themselves.
Since the path to the sign up form had to be made available
for non-authenticated users as well I have added a relaxation
constraint for the specified path.

I have also turned off autocomplete for the user and sign up
form since it seemed inappropriate that the login and password
fields were already filled out.

Also I have found a possible bug in Jetty 8. When the user is
already authenticated and accesses the login page it is handled
as if the user isn't authenticated. Hopefully the bug will be
fixed if I file it or if I provide a patch.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>